cache: update cache time when item is the same as was already cached
This is an incredibly stupid bug that I only noticed because my new-fangled
monitoring setup caused me to almost instantly hit an API ratelimit.

When we cached a post, we checked for if the already cached post matches
the one we just got, and if it does, we skip the cache operation entirely.
This is nice, since we save a set operation for post data we already have...
except it means that the last cache time is *not* updated, meaning that
every subsequent cache hit after the initial cache will fail, until
the post is updated again. This is fixed now.

(To be honest, the entire cache check for posts is flawed; posts will
update if the note count goes up, but we don't display it so that we can
cache renders for longer.)